+++ desc/objectives/__init__.py
@@ -1,7 +1,13 @@
 """Classes defining objectives for equilibrium and optimization."""
 
 from ._bootstrap import BootstrapRedlConsistency
-from ._coils import QuadraticFlux, SurfaceCurrentRegularization
+from ._coils import (
+    CoilCurvature,
+    CoilLength,
+    CoilTorsion,
+    QuadraticFlux,
+    SurfaceCurrentRegularization,
+)
 from ._equilibrium import (
     CurrentDensity,
     Energy,